Conditional logic for chartconfig specific setup and teardown logic

+++ integration/teardown/teardown.go
@@ -9,23 +9,28 @@ import (
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
 	"k8s.io/helm/pkg/helm"
+
+	"github.com/giantswarm/chart-operator/integration/env"
 )
 
 func Teardown(f *framework.Host, helmClient *helmclient.Client) error {
-	// clean host cluster components
+	// Clean operator components.
 	err := framework.HelmCmd("delete --purge giantswarm-chart-operator")
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	// clean guest cluster components
-	items := []string{"cnr-server", "apiextensions-chart-config-e2e"}
+	if env.TestedCustomResource() == env.ChartConfigCustomResource {
+		// Clean chartconfig related components.
+		items := []string{"cnr-server", "apiextensions-chart-config-e2e"}
 
-	for _, item := range items {
-		err := helmClient.DeleteRelease(context.TODO(), item, helm.DeletePurge(true))
-		if err != nil {
-			return microerror.Mask(err)
+		for _, item := range items {
+			err := helmClient.DeleteRelease(context.TODO(), item, helm.DeletePurge(true))
+			if err != nil {
+				return microerror.Mask(err)
+			}
 		}
 	}
+
 	return nil
 }
